Generated by Spark: it was actually the person card settings that i believed is messy, so have a go at that as per my previous prompt.
The actual card on the canvas actually became good. need the text to be a little bit bigger and remove the function that clicking on the image the image will appear. instead, double clicking anywhere in the card should open the card settings. also, make sure that the photo goes all the way up to the the top border of the card.

since the photo now in rectangular in the card, there should also be a rectangular photo in the card settings menu and it should be possible to adjust the left, right, up, down so that it will fit the borders optimal