Generated by Spark: When possible i would like that there will be no scrolling in the personal card settings menu. Maybe make the status colors a little bit smaller and mode advocate to the same line as importance.
Also, at the cars on the canvas, make sure that the advocate indicator is a star icon instead (and that it is yellow) and that it actually can be shown a little bit outdid the outer line. That should also be the fact for the stack icon, that could also ”overflow” the outer line of the card a little bit.